ARM: dts: nitrogen6_max: configure pullups/pulldowns on RS232/485 port
Configure pullups and pulldowns on GPIO pins connected to the ISL3330
RS-232/485 transceiver.

Note that these pins should properly be driven as outputs, but we
haven't yet ported the driver updates from the 3.0.35 kernel.

The state the pins should be in for RS-232 mode are as follows:
	PIN	GP	gpnum	state
	----	----	-----	-----
	485	2:24	56	low
	DEN	6:7	167	low
	ON	6:8	168	low
	RXEN	6:10	170	high

The following commands can be used to do this by hand:
	cd /sys/class/gpio
	for n in 56 167 168 170 ; do
		echo $n > export ;
		echo out > gpio$n/direction ;
	done

	echo -n 1 > gpio170/value
	for n in 56 167 168 ; do
		echo -n 0 > gpio$n/value ;
	done

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>